requirements.txt
altgraph==0.17.4
blinker==1.9.0
bottle==0.13.2
click==8.1.8
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
openpyxl==3.1.5
packaging==24.2
proxy_tools==0.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.16
pywebview==5.3.2
QtPy==2.4.2
setuptools==75.6.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
